requirements.txt
mlflow-skinny==2.20.0
accelerate==1.3.0
bitsandbytes==0.45.1
cmake==3.31.4
datasets==3.2.0
elasticsearch==8.17.1
exceptiongroup==1.2.2
faiss-cpu==1.9.0.post1
fastapi-cache2==0.2.2
ipykernel==6.29.5
langchain-community==0.3.15
langchain-huggingface==0.1.2
langchain-ollama==0.2.3
langchain-groq==0.2.4
langgraph==0.2.68
loguru==0.7.3
mlflow==2.20.0
pickleshare==0.7.5
pip-chill==1.0.3
sentencepiece==0.2.0
pandas
numpy
scikit-learn==1.6.1
sqlalchemy==2.0.38
pymysql==1.1.0
python-dotenv==1.0.1
pytest==7.4.0
pytest-cov==4.1.0
tqdm==4.67.1  # For progress bars
regex==2024.11.6  # Enhanced regex support
matplotlib==3.7.2
seaborn==0.12.2
pip-tools==7.3.0
PyMySQL==1.1.0
pycryptodome==3.20.0  # Provides Crypto.Cipher (AES)
PyJWT==2.8.0          # For JWT token handling
python-multipart==0.0.20